client/requirements.txt
requests
pyjks
ecdsa
pyopenssl